refactor into custom hook for data retrieval
